refactor(usdc): mainnet cutover via gov MsgUpdateParams, not a v34 migration
v34 executed on mainnet 2026-08-29, so its upgrade handler will never run
again — the AllowInjectiveUSDC wiring there was dead on arrival. The
allowlist is pure state, so no upgrade is needed at all: a governance
proposal carrying tokenization.MsgUpdateParams (full params object with
the canonical denom appended) does the mainnet job.

Drop the handler + its tests; keep the denom constants, DefaultParams,
genesis funding, and trace-derivation tests; add a test that exercises
the real gov path (authority check + full-replace semantics) against
pre-cutover state.
